feat(policy): register & implement class policy

+++ app/Http/Controllers/API/ClassController.php
@@ -20,6 +20,8 @@ class ClassController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny', IClass::class);
+
         return new ClassCollection(
             IClass::with('major')->search($request->search)->latest()->paginate($request->per_page)
         );
@@ -33,6 +35,8 @@ class ClassController extends Controller
      */
     public function store(StoreClassRequest $request)
     {
+        $this->authorize('create', IClass::class);
+
         return (new ClassResource(IClass::create($request->validated())->load('major')))
             ->additional(['message' => "Class has been Submitted successfully"]);
     }
@@ -45,6 +49,8 @@ class ClassController extends Controller
      */
     public function show(IClass $class)
     {
+        $this->authorize('view', $class);
+
         return new ClassResource($class->load('major'));
     }
 
@@ -57,6 +63,8 @@ class ClassController extends Controller
      */
     public function update(UpdateClassRequest $request, IClass $class)
     {
+        $this->authorize('update', $class);
+
         $class->update($request->validated());
 
         return (new ClassResource($class->load('major')))
@@ -71,6 +79,8 @@ class ClassController extends Controller
      */
     public function destroy(IClass $class)
     {
+        $this->authorize('delete', $class);
+
         $class->delete();
 
         return response(['message' => 'Class deleted successfully']);
